refactor: move `GrailsNameUtils` to utils package
