+++ gradle-plugin/src/main/groovy/grails/views/gradle/util/GrailsNameUtils.groovy
@@ -1,4 +1,4 @@
-package grails.views.gradle
+package grails.views.gradle.util
 
 import groovy.transform.CompileStatic
